Move UnifiedPush registration funcs to dedicated class
