+++ src/main/java/org/thoughtcrime/securesms/ConversationListActivity.java
@@ -69,6 +69,7 @@ import org.thoughtcrime.securesms.connect.DcHelper;
 import org.thoughtcrime.securesms.connect.DirectShareUtil;
 import org.thoughtcrime.securesms.geolocation.LocationStreamingService;
 import org.thoughtcrime.securesms.mms.GlideApp;
+import org.thoughtcrime.securesms.notifications.UnifiedPushUtils;
 import org.thoughtcrime.securesms.permissions.Permissions;
 import org.thoughtcrime.securesms.providers.PersistentBlobProvider;
 import org.thoughtcrime.securesms.proxy.ProxySettingsActivity;
@@ -86,8 +87,6 @@ import org.thoughtcrime.securesms.util.ShareUtil;
 import org.thoughtcrime.securesms.util.StorageUtil;
 import org.thoughtcrime.securesms.util.Util;
 import org.thoughtcrime.securesms.util.ViewUtil;
-import org.unifiedpush.android.connector.UnifiedPush;
-import org.unifiedpush.android.connector.data.ResolvedDistributor;
 
 public class ConversationListActivity extends PassphraseRequiredActionBarActivity
     implements ConversationListFragment.ConversationSelectedListener {
@@ -258,7 +257,7 @@ public class ConversationListActivity extends PassphraseRequiredActionBarActivit
       qrCodeHandler.secureJoinByQr(rawQrString, SecurejoinSource.Scan, SecurejoinUiPath.Unknown);
     }
 
-    mayInitUnifiedPush();
+    UnifiedPushUtils.mayInitUnifiedPush(this);
   }
 
   /**
@@ -328,82 +327,6 @@ public class ConversationListActivity extends PassphraseRequiredActionBarActivit
     }
 
     if (!wrongArch) Prefs.setBooleanPreference(this, NDK_ARCH_WARNED, false);
-  }
-
-  private void mayInitUnifiedPush() {
-    if (Prefs.isFcmPushEnabled(this)) {
-      // Do nothing, the application supports FCM
-      return;
-    }
-    if (!Prefs.unifiedPush(this)) {
-      // return if UnifiedPush is explicitly disabled
-      return;
-    }
-    if (UnifiedPush.getAckDistributor(this) != null) {
-      // Do nothing, UnifiedPush is initialized with ApplicationContext
-      return;
-    }
-    ResolvedDistributor resolvedDistributor = UnifiedPush.INSTANCE.resolveDefaultDistributor(this);
-    if (resolvedDistributor instanceof ResolvedDistributor.Found) {
-      // We now have a default distributor -> we use it
-      UnifiedPush.saveDistributor(
-          this, ((ResolvedDistributor.Found) resolvedDistributor).getPackageName());
-      ApplicationContext.getInstance(this).initializePush();
-    } else if (resolvedDistributor instanceof ResolvedDistributor.ToSelect) {
-      selectUnifiedPushDistributor();
-    }
-    // Else do nothing: the periodic sync is already setup during ApplicationContext init
-  }
-
-  /**
-   * The user has many distributors installed on the system, and none of them is defined as the
-   * default one: it needs to be selected with the OS dialog
-   *
-   * <p>To avoid the OS dialog to pop from nowhere, we introduce it with a simple dialog to show
-   * "You're about to select your UnifiedPush service." The OS dialog has a background showing
-   * "UnifiedPush" but some users may not see it. It is better to introduce with plain text.</p>
-   *
-   * <p>Note: The user necessarily knows about UnifiedPush: they have installed and enabled UnifiedPush
-   * on at least 2 distributors.</p>
-   */
-  private void selectUnifiedPushDistributor() {
-    DialogCallback callback = new DialogCallback() {
-      private Activity context = ConversationListActivity.this;
-      @Override
-      public void onCancel() {
-        Prefs.setUnifiedPush(context, false);
-      }
-
-      @Override
-      public void onConfirm() {
-        UnifiedPush.tryUseDefaultDistributor(context, success -> {
-          if (success) {
-            Prefs.resetReliableService(context);
-            ApplicationContext.getInstance(context).initializePush();
-          } else {
-            // The user has closed the OS dialog, we consider they don't want UnifiedPush
-            Prefs.setUnifiedPush(context, false);
-          }
-          return null;
-        });
-      }
-    };
-
-    introduceUnifiedPushDialog(callback);
-  }
-
-  private interface DialogCallback{
-    void onCancel();
-    void onConfirm();
-  }
-
-  private void introduceUnifiedPushDialog(DialogCallback callback) {
-    new AlertDialog.Builder(this)
-      .setMessage(R.string.dialog_introduce_unifiedpush_selection)
-      .setCancelable(true)
-      .setNegativeButton(android.R.string.cancel, (_d, _i) -> callback.onCancel())
-      .setPositiveButton(android.R.string.ok, (_d, _i) -> callback.onConfirm())
-      .show();
   }
 
   @Override
